Change parameters, small error to fix

+++ cpp/models/abm/world.cpp
@@ -157,10 +157,10 @@ Location& World::find_location(LocationType type, const Person& person)
     return get_individualized_location({index, type});
 }
 
-int World::get_subpopulation_combined(TimePoint t, InfectionState s, LocationType type) const
+size_t World::get_subpopulation_combined(TimePoint t, InfectionState s, LocationType type) const
 {
     auto& locs = m_locations[(uint32_t)type];
-    return std::accumulate(locs.begin(), locs.end(), 0, [&](int running_sum, const std::unique_ptr<Location>& loc) {
+    return std::accumulate(locs.begin(), locs.end(), 0, [&](size_t running_sum, const std::unique_ptr<Location>& loc) {
         return running_sum + loc->get_subpopulation(t, s);
     });
 }
